♻️ refactor(server): split ConfigService god-impl by concern
Move ~50 methods of the 2432-line impl into 7 topical submodules; config_service.rs prod drops to 858 lines, tests unchanged.
