+++ crates/awaken-server/src/services/config_service.rs
@@ -1,27 +1,33 @@
-use std::collections::HashSet;
 use std::sync::Arc;
-use std::time::Instant;
 
 use awaken_contract::AuditAction;
 use awaken_contract::contract::config_store::ConfigStore;
 use awaken_contract::contract::storage::StorageError;
 use awaken_contract::{
-    AgentSpec, ConfigRecord, McpServerSpec, ModelBindingSpec, ProviderSpec, RecordMeta,
-    RecordSource, ToolSpec, ToolSpecPatch, now_ms,
-};
-use awaken_runtime::registry::{
-    ProviderRemovalPreview, SerializableRegistryDiagnostic, diagnose_registry_set_serializable,
+    AgentSpec, ConfigRecord, McpServerSpec, ModelBindingSpec, ProviderSpec, ToolSpec,
 };
 use axum::http::HeaderMap;
-use serde_json::{Map, Value, json};
+use serde_json::{Value, json};
 
 use crate::app::AppState;
 use crate::services::audit_log::AuditLogger;
-use crate::services::config_envelope::{apply_overrides, unwrap_spec};
+use crate::services::config_envelope::unwrap_spec;
 
 use super::config_runtime::ConfigRuntimeError;
 
-const TOOLS_NAMESPACE: &str = "tools";
+mod agent_overrides;
+mod audit;
+mod mcp;
+mod normalization;
+mod provider;
+mod storage;
+mod tool_overrides;
+
+use normalization::{
+    classify_tool_source, effective_spec, effective_tool_spec, effective_visible_record,
+};
+
+pub(super) const TOOLS_NAMESPACE: &str = "tools";
 const OVERRIDES_NOT_SUPPORTED_FOR_USER_RECORD: &str =
     "overrides are not supported for user-source records; use PUT to update";
 
@@ -123,7 +129,7 @@ fn blocked_by_dependents(used_by: Vec<DependentRef>) -> ConfigServiceError {
     ))
 }
 
-fn overrides_not_supported_for_user_record() -> ConfigServiceError {
+pub(super) fn overrides_not_supported_for_user_record() -> ConfigServiceError {
     ConfigServiceError::InvalidPayload(OVERRIDES_NOT_SUPPORTED_FOR_USER_RECORD.into())
 }
 
@@ -392,67 +398,6 @@ impl<'a> ConfigService<'a> {
             out.push((id, meta));
         }
         Ok(out)
-    }
-
-    pub async fn preview_remove_provider(
-        &self,
-        id: &str,
-    ) -> Result<ProviderRemovalPreview, ConfigServiceError> {
-        if self
-            .store
-            .get(ConfigNamespace::Providers.as_str(), id)
-            .await?
-            .is_none()
-        {
-            return Err(ConfigServiceError::NotFound(format!("providers/{id}")));
-        }
-
-        let model_refs = self.find_dependents(ConfigNamespace::Providers, id).await?;
-        let model_ids = model_refs
-            .into_iter()
-            .map(|reference| reference.id)
-            .collect::<Vec<_>>();
-        let agent_ids = self.agents_referencing_models(&model_ids).await?;
-
-        Ok(ProviderRemovalPreview::new(id, model_ids, agent_ids))
-    }
-
-    /// Single-pass scan: returns agent ids whose effective spec references
-    /// any of the supplied `model_ids` and does not declare a remote endpoint.
-    /// Avoids the previous O(model_ids × agents) double scan in
-    /// `preview_remove_provider` and the provider force-cascade path.
-    async fn agents_referencing_models(
-        &self,
-        model_ids: &[String],
-    ) -> Result<Vec<String>, ConfigServiceError> {
-        if model_ids.is_empty() {
-            return Ok(Vec::new());
-        }
-        let model_id_set: HashSet<&str> = model_ids.iter().map(String::as_str).collect();
-        let agents = self.store.list("agents", 0, usize::MAX).await?;
-        let mut refs = Vec::new();
-        for (agent_id, value) in agents {
-            let Some(agent) = effective_visible_record::<AgentSpec>(value)? else {
-                continue;
-            };
-            if agent.endpoint.is_none() && model_id_set.contains(agent.model_id.as_str()) {
-                refs.push(agent_id);
-            }
-        }
-        Ok(refs)
-    }
-
-    pub fn registry_diagnostics(
-        &self,
-    ) -> Result<Vec<SerializableRegistryDiagnostic>, ConfigServiceError> {
-        let registries = self
-            .state
-            .runtime
-            .registry_set()
-            .ok_or(ConfigServiceError::Apply(
-                "runtime does not expose a configurable registry snapshot".into(),
-            ))?;
-        Ok(diagnose_registry_set_serializable(&registries))
     }
 
     /// Dry-run validation. Runs the same `prepare_body` + `validate_payload`
@@ -896,1508 +841,9 @@ impl<'a> ConfigService<'a> {
             ConfigNamespace::Agents | ConfigNamespace::McpServers => Ok(vec![]),
         }
     }
-
-    /// Emit an audit event if an audit logger is configured.
-    ///
-    /// Best-effort: the call is fire-and-forget (matching the `AuditLogger::emit` contract).
-    async fn emit_audit(
-        &self,
-        action: AuditAction,
-        namespace: ConfigNamespace,
-        id: &str,
-        before: Option<Value>,
-        after: Option<Value>,
-        headers: &HeaderMap,
-    ) {
-        self.emit_audit_with_suffix(action, namespace, id, "", before, after, headers)
-            .await;
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn emit_audit_with_suffix(
-        &self,
-        action: AuditAction,
-        namespace: ConfigNamespace,
-        id: &str,
-        suffix: &str,
-        before: Option<Value>,
-        after: Option<Value>,
-        headers: &HeaderMap,
-    ) {
-        self.emit_audit_with_suffix_in_namespace(
-            action,
-            namespace.as_str(),
-            id,
-            suffix,
-            before,
-            after,
-            headers,
-        )
-        .await;
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn emit_audit_with_suffix_in_namespace(
-        &self,
-        action: AuditAction,
-        namespace: &str,
-        id: &str,
-        suffix: &str,
-        before: Option<Value>,
-        after: Option<Value>,
-        headers: &HeaderMap,
-    ) {
-        let Some(audit) = &self.audit else {
-            return;
-        };
-        let resource = if suffix.is_empty() {
-            format!("{namespace}/{id}")
-        } else {
-            format!("{namespace}/{id}/{suffix}")
-        };
-        audit.emit(action, &resource, before, after, headers).await;
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn emit_audit_apply_failed(
-        &self,
-        namespace: ConfigNamespace,
-        id: &str,
-        suffix: &str,
-        before: Option<Value>,
-        after: Option<Value>,
-        error_msg: String,
-        headers: &HeaderMap,
-    ) {
-        self.emit_audit_apply_failed_in_namespace(
-            namespace.as_str(),
-            id,
-            suffix,
-            before,
-            after,
-            error_msg,
-            headers,
-        )
-        .await;
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn emit_audit_apply_failed_in_namespace(
-        &self,
-        namespace: &str,
-        id: &str,
-        suffix: &str,
-        before: Option<Value>,
-        after: Option<Value>,
-        error_msg: String,
-        headers: &HeaderMap,
-    ) {
-        let Some(audit) = &self.audit else {
-            return;
-        };
-        let resource = if suffix.is_empty() {
-            format!("{namespace}/{id}")
-        } else {
-            format!("{namespace}/{id}/{suffix}")
-        };
-        audit
-            .emit_apply_failed(&resource, before, after, error_msg, headers)
-            .await;
-    }
-
-    fn runtime_manager(
-        &self,
-    ) -> Result<&Arc<crate::services::config_runtime::ConfigRuntimeManager>, ConfigServiceError>
-    {
-        self.state
-            .config_runtime_manager
-            .as_ref()
-            .ok_or(ConfigServiceError::NotEnabled)
-    }
-
-    fn user_record_from_body(body: &Value, previous: Option<&Value>) -> ConfigRecord<Value> {
-        let mut meta = RecordMeta::new_user();
-        if let Some(previous) = previous
-            && let Ok(previous_record) = ConfigRecord::<Value>::from_value(previous.clone())
-            && previous_record.meta.created_at != 0
-        {
-            meta.created_at = previous_record.meta.created_at;
-        }
-        ConfigRecord {
-            spec: body.clone(),
-            meta,
-        }
-    }
-
-    fn storage_write_error(
-        namespace: ConfigNamespace,
-        id: &str,
-        error: StorageError,
-    ) -> ConfigServiceError {
-        Self::storage_write_error_for_namespace(namespace.as_str(), id, error)
-    }
-
-    fn storage_write_error_for_namespace(
-        namespace: &str,
-        id: &str,
-        error: StorageError,
-    ) -> ConfigServiceError {
-        match error {
-            StorageError::AlreadyExists(_) => {
-                ConfigServiceError::Conflict(format!("{namespace}/{id} already exists"))
-            }
-            StorageError::VersionConflict { expected, actual } => {
-                ConfigServiceError::Conflict(format!(
-                    "{}/{} was modified by another writer (expected revision {expected}, found {actual}); retry the mutation",
-                    namespace, id,
-                ))
-            }
-            other => ConfigServiceError::Storage(other),
-        }
-    }
-
-    async fn insert_record_absent<T: serde::Serialize + serde::de::DeserializeOwned>(
-        &self,
-        namespace: ConfigNamespace,
-        id: &str,
-        record: &mut ConfigRecord<T>,
-        revision: u64,
-    ) -> Result<u64, ConfigServiceError> {
-        record.meta.revision = revision;
-        let envelope = record
-            .to_value()
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-        self.store
-            .put_if_absent(namespace.as_str(), id, &envelope)
-            .await
-            .map(|()| revision)
-            .map_err(|error| Self::storage_write_error(namespace, id, error))
-    }
-
-    /// Write `record` using `put_if_revision`, bumping `meta.revision` from
-    /// `expected_revision`. Returns the new revision on success or
-    /// `ConfigServiceError::Conflict` on CAS mismatch.
-    async fn cas_put_record<T: serde::Serialize + serde::de::DeserializeOwned>(
-        &self,
-        namespace: ConfigNamespace,
-        id: &str,
-        record: &mut ConfigRecord<T>,
-        expected_revision: u64,
-    ) -> Result<u64, ConfigServiceError> {
-        self.cas_put_record_in_namespace(namespace.as_str(), id, record, expected_revision)
-            .await
-    }
-
-    async fn cas_put_record_in_namespace<T: serde::Serialize + serde::de::DeserializeOwned>(
-        &self,
-        namespace: &str,
-        id: &str,
-        record: &mut ConfigRecord<T>,
-        expected_revision: u64,
-    ) -> Result<u64, ConfigServiceError> {
-        let next_revision = expected_revision.saturating_add(1);
-        record.meta.revision = next_revision;
-        let envelope = record
-            .to_value()
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-        self.store
-            .put_if_revision(namespace, id, &envelope, expected_revision)
-            .await
-            .map(|()| next_revision)
-            .map_err(|error| Self::storage_write_error_for_namespace(namespace, id, error))
-    }
-
-    async fn cas_delete_record(
-        &self,
-        namespace: ConfigNamespace,
-        id: &str,
-        expected_revision: u64,
-    ) -> Result<(), ConfigServiceError> {
-        self.store
-            .delete_if_revision(namespace.as_str(), id, expected_revision)
-            .await
-            .map_err(|error| Self::storage_write_error(namespace, id, error))
-    }
-
-    async fn rollback_to_raw_after_revision(
-        &self,
-        namespace: ConfigNamespace,
-        id: &str,
-        raw: Value,
-        expected_revision: u64,
-    ) -> Result<u64, ConfigServiceError> {
-        let mut rollback = ConfigRecord::<Value>::from_value(raw)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-        self.cas_put_record(namespace, id, &mut rollback, expected_revision)
-            .await
-    }
-
-    async fn rollback_to_raw_after_revision_in_namespace(
-        &self,
-        namespace: &str,
-        id: &str,
-        raw: Value,
-        expected_revision: u64,
-    ) -> Result<u64, ConfigServiceError> {
-        let mut rollback = ConfigRecord::<Value>::from_value(raw)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-        self.cas_put_record_in_namespace(namespace, id, &mut rollback, expected_revision)
-            .await
-    }
-
-    async fn rollback_deleted_records(
-        &self,
-        deleted_records: Vec<(ConfigNamespace, String, Value, u64)>,
-    ) -> Result<(), ConfigServiceError> {
-        for (rollback_namespace, rollback_id, raw, revision) in deleted_records.into_iter().rev() {
-            let mut rollback = ConfigRecord::<Value>::from_value(raw)
-                .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-            self.insert_record_absent(
-                rollback_namespace,
-                &rollback_id,
-                &mut rollback,
-                revision + 1,
-            )
-            .await?;
-        }
-        Ok(())
-    }
-
-    async fn persist_and_apply_locked(
-        &self,
-        manager: &crate::services::config_runtime::ConfigRuntimeManager,
-        namespace: ConfigNamespace,
-        id: &str,
-        previous: Option<Value>,
-        body: Value,
-        headers: &HeaderMap,
-    ) -> Result<Value, ConfigServiceError> {
-        self.validate_payload(namespace, &body)?;
-        let mut record = Self::user_record_from_body(&body, previous.as_ref());
-        let write_revision = match previous.as_ref() {
-            Some(previous) => {
-                let expected_revision = ConfigRecord::<Value>::from_value(previous.clone())
-                    .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?
-                    .meta
-                    .revision;
-                self.cas_put_record(namespace, id, &mut record, expected_revision)
-                    .await?
-            }
-            None => {
-                self.insert_record_absent(namespace, id, &mut record, 1)
-                    .await?
-            }
-        };
-
-        let apply_result = manager
-            .apply_locked()
-            .await
-            .map(|_| ())
-            .map_err(map_runtime_error);
-        if let Err(error) = apply_result {
-            self.emit_audit_apply_failed(
-                namespace,
-                id,
-                "",
-                previous.as_ref().map(|p| unwrap_spec(p.clone())),
-                Some(unwrap_spec(body.clone())),
-                error.to_string(),
-                headers,
-            )
-            .await;
-            match previous {
-                Some(previous) => {
-                    self.rollback_to_raw_after_revision(namespace, id, previous, write_revision)
-                        .await?;
-                }
-                None => {
-                    self.cas_delete_record(namespace, id, write_revision)
-                        .await?
-                }
-            }
-            return Err(error);
-        }
-
-        self.redact_response(namespace, body)
-    }
-
-    async fn prepare_body(
-        &self,
-        namespace: ConfigNamespace,
-        path_id: Option<&str>,
-        body: Value,
-    ) -> Result<(String, Value), ConfigServiceError> {
-        let mut object = into_object(body)?;
-        let id = object
-            .get("id")
-            .and_then(Value::as_str)
-            .map(str::trim)
-            .filter(|id| !id.is_empty())
-            .ok_or(ConfigServiceError::MissingId)?
-            .to_string();
-
-        if let Some(path_id) = path_id
-            && path_id != id
-        {
-            return Err(ConfigServiceError::InvalidPayload(format!(
-                "path id '{path_id}' does not match body id '{id}'"
-            )));
-        }
-
-        match namespace {
-            ConfigNamespace::Providers => {
-                object.remove("has_api_key");
-                self.normalize_provider_payload(path_id, &mut object)
-                    .await?;
-            }
-            ConfigNamespace::McpServers => {
-                object.remove("has_env");
-                object.remove("env_keys");
-                self.normalize_mcp_server_payload(path_id, &mut object)
-                    .await?;
-            }
-            ConfigNamespace::Agents | ConfigNamespace::Models => {}
-        }
-
-        object.remove("created_at");
-        object.remove("updated_at");
-
-        Ok((id, Value::Object(object)))
-    }
-
-    async fn normalize_provider_payload(
-        &self,
-        path_id: Option<&str>,
-        body: &mut Map<String, Value>,
-    ) -> Result<(), ConfigServiceError> {
-        let explicit_clear = matches!(body.get("api_key"), Some(Value::String(value)) if value.is_empty())
-            || matches!(body.get("api_key"), Some(Value::Null));
-        if explicit_clear {
-            body.remove("api_key");
-            return Ok(());
-        }
-
-        if body.contains_key("api_key") || path_id.is_none() {
-            return Ok(());
-        }
-
-        let Some(path_id) = path_id else {
-            return Ok(());
-        };
-        let Some(existing) = self
-            .store
-            .get(ConfigNamespace::Providers.as_str(), path_id)
-            .await?
-        else {
-            return Ok(());
-        };
-        // The stored value may be either a bare spec or a ConfigRecord envelope.
-        // Navigate into spec if needed before accessing fields.
-        let spec_value = unwrap_spec(existing);
-        let Some(existing_object) = spec_value.as_object() else {
-            return Ok(());
-        };
-        if let Some(existing_key) = existing_object.get("api_key") {
-            body.insert("api_key".into(), existing_key.clone());
-        }
-        Ok(())
-    }
-
-    fn validate_payload(
-        &self,
-        namespace: ConfigNamespace,
-        body: &Value,
-    ) -> Result<(), ConfigServiceError> {
-        match namespace {
-            ConfigNamespace::Agents => {
-                awaken_contract::validate_agent_spec(body.clone())
-                    .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-            }
-            ConfigNamespace::Models => {
-                awaken_contract::validate_model_binding_spec(body.clone())
-                    .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-            }
-            ConfigNamespace::Providers => {
-                let spec = awaken_contract::validate_provider_spec(body.clone())
-                    .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-                // Eager credential validation: parse `credentials_kind` and the
-                // (kind × adapter × api_key) shape so misconfigured providers
-                // are rejected at write time, not at first inference. The
-                // adapter string is **not** validated here because the provider
-                // may be saved before its target adapter is rolled out (admin
-                // UI accepts unknown adapter names with a server-side error
-                // surface). Adapter-specific validation lives in the runtime
-                // build path.
-                let kind = awaken_runtime::credentials::CredentialKind::from_options(
-                    &spec.adapter_options,
-                )
-                .map_err(ConfigServiceError::InvalidPayload)?;
-                awaken_runtime::credentials::build_material(
-                    &spec.adapter,
-                    kind,
-                    spec.api_key.as_ref(),
-                )
-                .map_err(ConfigServiceError::InvalidPayload)?;
-            }
-            ConfigNamespace::McpServers => {
-                let spec: McpServerSpec = from_value(body)?;
-                if spec.id.trim().is_empty() {
-                    return Err(ConfigServiceError::InvalidPayload(
-                        "mcp server id cannot be empty".into(),
-                    ));
-                }
-
-                match spec.transport {
-                    awaken_contract::McpTransportKind::Stdio => {
-                        if spec
-                            .command
-                            .as_deref()
-                            .is_none_or(|value| value.trim().is_empty())
-                        {
-                            return Err(ConfigServiceError::InvalidPayload(
-                                "stdio mcp server requires a non-empty command".into(),
-                            ));
-                        }
-                    }
-                    awaken_contract::McpTransportKind::Http => {
-                        if spec
-                            .url
-                            .as_deref()
-                            .is_none_or(|value| value.trim().is_empty())
-                        {
-                            return Err(ConfigServiceError::InvalidPayload(
-                                "http mcp server requires a non-empty url".into(),
-                            ));
-                        }
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    fn redact_response(
-        &self,
-        namespace: ConfigNamespace,
-        value: Value,
-    ) -> Result<Value, ConfigServiceError> {
-        match namespace {
-            ConfigNamespace::Providers => {
-                let mut object = into_object(value)?;
-                let has_api_key = object
-                    .get("api_key")
-                    .and_then(Value::as_str)
-                    .is_some_and(|value| !value.is_empty());
-                object.remove("api_key");
-                if has_api_key {
-                    object.insert("has_api_key".into(), Value::Bool(true));
-                }
-                Ok(Value::Object(object))
-            }
-            ConfigNamespace::McpServers => {
-                let mut object = into_object(value)?;
-                let env_keys = object
-                    .get("env")
-                    .and_then(Value::as_object)
-                    .map(|env| {
-                        let mut keys = env.keys().cloned().collect::<Vec<_>>();
-                        keys.sort();
-                        keys
-                    })
-                    .unwrap_or_default();
-                object.remove("env");
-                if !env_keys.is_empty() {
-                    object.insert("has_env".into(), Value::Bool(true));
-                    object.insert(
-                        "env_keys".into(),
-                        Value::Array(env_keys.into_iter().map(Value::String).collect()),
-                    );
-                }
-                Ok(Value::Object(object))
-            }
-            ConfigNamespace::Agents | ConfigNamespace::Models => Ok(value),
-        }
-    }
-
-    /// Test whether a stored provider config is usable.
-    ///
-    /// Strategy depends on `credentials_kind`:
-    ///
-    /// - **Static / bearer** (default): construction-only probe. Loads the
-    ///   stored `ProviderSpec` and runs `build_genai_provider_executor` to
-    ///   prove that the adapter name parses, the api_key (if any) is the
-    ///   right shape, and adapter_options are valid. **No network call.**
-    ///
-    /// - **Dynamic** (`service_account_json`, future cloud creds):
-    ///   construction-only probe **plus** a live token mint via the
-    ///   credential broker. This catches revoked keys, deleted service
-    ///   accounts, unreachable token endpoints, and missing scopes —
-    ///   problems that a construction probe cannot see. The mint reuses
-    ///   the same broker code that production inference does, so a
-    ///   passing test is strong evidence that the next inference will
-    ///   succeed at the auth layer.
-    ///
-    /// In both cases the LLM endpoint itself is not contacted; that
-    /// would require a full runtime context (cancellation, streaming,
-    /// observability) and would also bill the user for a token.
-    pub async fn test_provider(&self, id: &str) -> Result<ProviderTestResult, ConfigServiceError> {
-        let raw = self
-            .store
-            .get(ConfigNamespace::Providers.as_str(), id)
-            .await?
-            .ok_or_else(|| ConfigServiceError::NotFound(format!("providers/{id}")))?;
-
-        let spec: ProviderSpec = ConfigRecord::<ProviderSpec>::from_value(raw)
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))
-            .map(|r| r.spec)?;
-
-        // Construction probe: catches adapter parsing, material parsing,
-        // header validation, and any other build-time check. Reuses the
-        // production builder so any change to the build path is covered
-        // here automatically.
-        let start = Instant::now();
-        let broker: std::sync::Arc<dyn awaken_runtime::credentials::CredentialBroker> =
-            std::sync::Arc::new(awaken_runtime::credentials::AwakenCredentialBroker::new());
-        let build_result =
-            crate::services::config_runtime::build_genai_provider_executor_with_broker(
-                &spec,
-                std::sync::Arc::clone(&broker),
-            );
-        let mut latency_ms = start.elapsed().as_millis() as u64;
-
-        if let Err(e) = build_result {
-            return Ok(ProviderTestResult {
-                ok: false,
-                latency_ms,
-                network_tested: false,
-                error: Some(redact_provider_error(&e.to_string(), &spec)),
-            });
-        }
-
-        // Pre-flight token mint for dynamic credentials. Skipped for bearer
-        // (static / env-var fallback) where the broker would either no-op
-        // or hand back the static value — neither tests anything new.
-        let kind = match awaken_runtime::credentials::CredentialKind::from_options(
-            &spec.adapter_options,
-        ) {
-            Ok(k) => k,
-            Err(_) => {
-                // Already caught by the build probe above; defensive-coded.
-                return Ok(ProviderTestResult {
-                    ok: true,
-                    latency_ms,
-                    network_tested: false,
-                    error: None,
-                });
-            }
-        };
-        let mut network_tested = false;
-        if matches!(
-            kind,
-            awaken_runtime::credentials::CredentialKind::GoogleServiceAccountJson
-        ) {
-            let scope = "https://www.googleapis.com/auth/cloud-platform";
-            let mint_start = Instant::now();
-            network_tested = true;
-            let mint_result = broker.token_for(&spec.id, scope).await;
-            latency_ms = latency_ms.saturating_add(mint_start.elapsed().as_millis() as u64);
-            if let Err(err) = mint_result {
-                return Ok(ProviderTestResult {
-                    ok: false,
-                    latency_ms,
-                    network_tested,
-                    error: Some(redact_provider_error(&err.to_string(), &spec)),
-                });
-            }
-        }
-
-        Ok(ProviderTestResult {
-            ok: true,
-            latency_ms,
-            network_tested,
-            error: None,
-        })
-    }
-
-    async fn normalize_mcp_server_payload(
-        &self,
-        path_id: Option<&str>,
-        body: &mut Map<String, Value>,
-    ) -> Result<(), ConfigServiceError> {
-        if body.contains_key("env") || path_id.is_none() {
-            return Ok(());
-        }
-
-        let Some(path_id) = path_id else {
-            return Ok(());
-        };
-        let Some(existing) = self
-            .store
-            .get(ConfigNamespace::McpServers.as_str(), path_id)
-            .await?
-        else {
-            return Ok(());
-        };
-        // The stored value may be either a bare spec or a ConfigRecord envelope.
-        let spec_value = unwrap_spec(existing);
-        let Some(existing_object) = spec_value.as_object() else {
-            return Ok(());
-        };
-        if let Some(existing_env) = existing_object.get("env") {
-            body.insert("env".into(), existing_env.clone());
-        }
-        Ok(())
-    }
-
-    /// PATCH /v1/config/agents/:id/overrides
-    ///
-    /// Merges the patch body into the existing `user_overrides` of a Builtin
-    /// agent record. JSON null clears nullable AgentSpec fields; for other
-    /// fields it removes the existing override. Non-null keys overwrite.
-    /// Returns the effective AgentSpec after the merge.
-    pub async fn patch_agent_overrides(
-        &self,
-        id: &str,
-        body: Value,
-        headers: &HeaderMap,
-    ) -> Result<Value, ConfigServiceError> {
-        let manager = self.runtime_manager()?;
-        let _apply_guard = manager.lock_apply().await;
-
-        let raw = self
-            .store
-            .get(ConfigNamespace::Agents.as_str(), id)
-            .await?
-            .ok_or_else(|| ConfigServiceError::NotFound(format!("agents/{id}")))?;
-
-        let mut record = ConfigRecord::<AgentSpec>::from_value(raw.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        if matches!(record.meta.source, RecordSource::User) {
-            return Err(overrides_not_supported_for_user_record());
-        }
-
-        let expected_revision = record.meta.revision;
-
-        // Validate incoming body field names via AgentSpecPatch (deny_unknown_fields).
-        // We use a separate check for null values: replace nulls with a dummy value
-        // so that deny_unknown_fields can still catch unknown field names.
-        let body_map = match &body {
-            Value::Object(m) => m,
-            _ => {
-                return Err(ConfigServiceError::InvalidPayload(
-                    "expected JSON object body".into(),
-                ));
-            }
-        };
-        // Null values for nullable AgentSpec fields are valid (they mean
-        // "clear the base value").
-        // Pass the body as-is; deny_unknown_fields catches unknown field names.
-        awaken_contract::validate_agent_spec_patch(body.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        // Merge patch INTO existing user_overrides (shallow key-level merge).
-        // Use the raw body Value to preserve nullable-field nulls.
-        let mut existing_map: Map<String, Value> = record
-            .meta
-            .user_overrides
-            .as_ref()
-            .and_then(Value::as_object)
-            .cloned()
-            .unwrap_or_default();
-
-        for (k, v) in body_map {
-            if v.is_null() {
-                if is_nullable_agent_patch_field(k) {
-                    existing_map.insert(k.clone(), Value::Null);
-                } else {
-                    existing_map.remove(k);
-                }
-            } else {
-                existing_map.insert(k.clone(), v.clone());
-            }
-        }
-
-        // Validate the merged overrides by round-tripping through AgentSpecPatch.
-        let merged_value = Value::Object(existing_map.clone());
-        awaken_contract::validate_agent_spec_patch(merged_value.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        let proposed_overrides: Option<Value> = if existing_map.is_empty() {
-            None
-        } else {
-            Some(merged_value.clone())
-        };
-
-        // Short-circuit: if the proposed overrides are identical to existing ones,
-        // skip the store write, apply_locked, and audit emit — it's a no-op.
-        if proposed_overrides == record.meta.user_overrides {
-            let effective_spec = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-                .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-            return serde_json::to_value(&effective_spec)
-                .map_err(|e| ConfigServiceError::Serialization(e.to_string()));
-        }
-
-        let before_spec = apply_overrides(record.spec.clone(), record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let before = serde_json::to_value(&before_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        record.meta.user_overrides = proposed_overrides;
-        record.meta.updated_at = now_ms();
-
-        let write_revision = self
-            .cas_put_record(ConfigNamespace::Agents, id, &mut record, expected_revision)
-            .await?;
-
-        let apply_result = manager
-            .apply_locked()
-            .await
-            .map(|_| ())
-            .map_err(map_runtime_error);
-        if let Err(error) = apply_result {
-            self.emit_audit_apply_failed(
-                ConfigNamespace::Agents,
-                id,
-                "overrides",
-                Some(before.clone()),
-                None,
-                error.to_string(),
-                headers,
-            )
-            .await;
-            self.rollback_to_raw_after_revision(ConfigNamespace::Agents, id, raw, write_revision)
-                .await?;
-            return Err(error);
-        }
-
-        let after_spec = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let after = serde_json::to_value(&after_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        self.emit_audit_with_suffix(
-            AuditAction::Update,
-            ConfigNamespace::Agents,
-            id,
-            "overrides",
-            Some(before),
-            Some(after.clone()),
-            headers,
-        )
-        .await;
-
-        Ok(after)
-    }
-
-    /// DELETE /v1/config/agents/:id/overrides
-    ///
-    /// Clears all user overrides from a Builtin agent record. Returns the
-    /// effective AgentSpec (which is now the bare base spec, no overrides).
-    pub async fn clear_agent_overrides(
-        &self,
-        id: &str,
-        headers: &HeaderMap,
-    ) -> Result<Value, ConfigServiceError> {
-        let manager = self.runtime_manager()?;
-        let _apply_guard = manager.lock_apply().await;
-
-        let raw = self
-            .store
-            .get(ConfigNamespace::Agents.as_str(), id)
-            .await?
-            .ok_or_else(|| ConfigServiceError::NotFound(format!("agents/{id}")))?;
-
-        let mut record = ConfigRecord::<AgentSpec>::from_value(raw.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        if matches!(record.meta.source, RecordSource::User) {
-            return Err(overrides_not_supported_for_user_record());
-        }
-
-        // Short-circuit: if overrides are already None, this is a no-op — skip
-        // the store write, apply_locked, and audit emit.
-        if record.meta.user_overrides.is_none() {
-            return serde_json::to_value(&record.spec)
-                .map_err(|e| ConfigServiceError::Serialization(e.to_string()));
-        }
-
-        let expected_revision = record.meta.revision;
-
-        let before_spec = apply_overrides(record.spec.clone(), record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let before = serde_json::to_value(&before_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        record.meta.user_overrides = None;
-        record.meta.updated_at = now_ms();
-
-        let write_revision = self
-            .cas_put_record(ConfigNamespace::Agents, id, &mut record, expected_revision)
-            .await?;
-
-        let apply_result = manager
-            .apply_locked()
-            .await
-            .map(|_| ())
-            .map_err(map_runtime_error);
-        if let Err(error) = apply_result {
-            self.emit_audit_apply_failed(
-                ConfigNamespace::Agents,
-                id,
-                "overrides",
-                Some(before.clone()),
-                None,
-                error.to_string(),
-                headers,
-            )
-            .await;
-            self.rollback_to_raw_after_revision(ConfigNamespace::Agents, id, raw, write_revision)
-                .await?;
-            return Err(error);
-        }
-
-        let after = serde_json::to_value(&record.spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        self.emit_audit_with_suffix(
-            AuditAction::Update,
-            ConfigNamespace::Agents,
-            id,
-            "overrides",
-            Some(before),
-            Some(after.clone()),
-            headers,
-        )
-        .await;
-
-        Ok(after)
-    }
-
-    /// DELETE /v1/config/agents/:id/overrides/:field
-    ///
-    /// Removes a single field from the user overrides of a Builtin agent record.
-    /// Returns 400 if `field` is not a recognized AgentSpecPatch field.
-    /// Idempotent: if the field is not present in user_overrides, returns the
-    /// current effective spec without writing to the store or emitting an audit event.
-    pub async fn clear_agent_override_field(
-        &self,
-        id: &str,
-        field: &str,
-        headers: &HeaderMap,
-    ) -> Result<Value, ConfigServiceError> {
-        let manager = self.runtime_manager()?;
-        let _apply_guard = manager.lock_apply().await;
-
-        let raw = self
-            .store
-            .get(ConfigNamespace::Agents.as_str(), id)
-            .await?
-            .ok_or_else(|| ConfigServiceError::NotFound(format!("agents/{id}")))?;
-
-        let mut record = ConfigRecord::<AgentSpec>::from_value(raw.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        if matches!(record.meta.source, RecordSource::User) {
-            return Err(overrides_not_supported_for_user_record());
-        }
-
-        let expected_revision = record.meta.revision;
-
-        let before_spec = apply_overrides(record.spec.clone(), record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let before = serde_json::to_value(&before_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        // Validate that field is recognized by AgentSpecPatch before mutating.
-        // Use a null probe: `AgentSpecPatch` accepts null for all Option fields, and
-        // deny_unknown_fields will reject unknown field names.
-        let probe = Value::Object({
-            let mut m = Map::new();
-            m.insert(field.to_string(), Value::Null);
-            m
-        });
-        awaken_contract::validate_agent_spec_patch(probe).map_err(|_| {
-            ConfigServiceError::InvalidPayload(format!("unknown override field: {field}"))
-        })?;
-
-        // Remove the field from existing overrides.
-        let mut existing_map: Map<String, Value> = record
-            .meta
-            .user_overrides
-            .as_ref()
-            .and_then(Value::as_object)
-            .cloned()
-            .unwrap_or_default();
-
-        // Short-circuit: if the field is not present in overrides, this is a no-op —
-        // skip the store write, apply_locked, and audit emit.
-        if !existing_map.contains_key(field) {
-            let effective_spec = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-                .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-            return serde_json::to_value(&effective_spec)
-                .map_err(|e| ConfigServiceError::Serialization(e.to_string()));
-        }
-
-        existing_map.remove(field);
-
-        let merged_value = Value::Object(existing_map.clone());
-        record.meta.user_overrides = if existing_map.is_empty() {
-            None
-        } else {
-            Some(merged_value)
-        };
-        record.meta.updated_at = now_ms();
-
-        let write_revision = self
-            .cas_put_record(ConfigNamespace::Agents, id, &mut record, expected_revision)
-            .await?;
-
-        let apply_result = manager
-            .apply_locked()
-            .await
-            .map(|_| ())
-            .map_err(map_runtime_error);
-        if let Err(error) = apply_result {
-            self.emit_audit_apply_failed(
-                ConfigNamespace::Agents,
-                id,
-                &format!("overrides/{field}"),
-                Some(before.clone()),
-                None,
-                error.to_string(),
-                headers,
-            )
-            .await;
-            self.rollback_to_raw_after_revision(ConfigNamespace::Agents, id, raw, write_revision)
-                .await?;
-            return Err(error);
-        }
-
-        let after_spec = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let after = serde_json::to_value(&after_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        self.emit_audit_with_suffix(
-            AuditAction::Update,
-            ConfigNamespace::Agents,
-            id,
-            &format!("overrides/{field}"),
-            Some(before),
-            Some(after.clone()),
-            headers,
-        )
-        .await;
-
-        Ok(after)
-    }
-
-    /// PATCH /v1/config/tools/:id/overrides — see ADR-0029.
-    pub async fn patch_tool_overrides(
-        &self,
-        id: &str,
-        body: Value,
-        headers: &HeaderMap,
-    ) -> Result<Value, ConfigServiceError> {
-        const MAX_DESCRIPTION_LEN: usize = 4096;
-
-        let manager = self.runtime_manager()?;
-        let _apply_guard = manager.lock_apply().await;
-
-        let raw = self
-            .store
-            .get(TOOLS_NAMESPACE, id)
-            .await?
-            .ok_or_else(|| ConfigServiceError::NotFound(format!("tools/{id}")))?;
-
-        let mut record = ConfigRecord::<ToolSpec>::from_value(raw.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        if matches!(record.meta.source, RecordSource::User) {
-            return Err(overrides_not_supported_for_user_record());
-        }
-
-        let expected_revision = record.meta.revision;
-
-        let body_map = match &body {
-            Value::Object(m) => m,
-            _ => {
-                return Err(ConfigServiceError::InvalidPayload(
-                    "expected JSON object body".into(),
-                ));
-            }
-        };
-
-        // Schema validation: deny_unknown_fields catches bad field names.
-        let _: ToolSpecPatch = serde_json::from_value(body.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        // Value validation (non-empty trim, length cap).
-        if let Some(Value::String(s)) = body_map.get("description") {
-            let trimmed = s.trim();
-            if trimmed.is_empty() {
-                return Err(ConfigServiceError::InvalidPayload(
-                    "description must be non-empty".into(),
-                ));
-            }
-            if s.len() > MAX_DESCRIPTION_LEN {
-                return Err(ConfigServiceError::InvalidPayload(format!(
-                    "description exceeds {MAX_DESCRIPTION_LEN}-byte limit"
-                )));
-            }
-        }
-
-        // Shallow merge into existing user_overrides; null = clear.
-        let mut existing_map: Map<String, Value> = record
-            .meta
-            .user_overrides
-            .as_ref()
-            .and_then(Value::as_object)
-            .cloned()
-            .unwrap_or_default();
-        for (k, v) in body_map {
-            if v.is_null() {
-                existing_map.remove(k);
-            } else {
-                existing_map.insert(k.clone(), v.clone());
-            }
-        }
-
-        // Re-validate the merged overrides shape.
-        let merged_value = Value::Object(existing_map.clone());
-        let _: ToolSpecPatch = serde_json::from_value(merged_value.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        let proposed_overrides: Option<Value> = if existing_map.is_empty() {
-            None
-        } else {
-            Some(merged_value.clone())
-        };
-
-        if proposed_overrides == record.meta.user_overrides {
-            let effective_spec = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-                .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-            return serde_json::to_value(&effective_spec)
-                .map_err(|e| ConfigServiceError::Serialization(e.to_string()));
-        }
-
-        let before_spec = apply_overrides(record.spec.clone(), record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let before = serde_json::to_value(&before_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        record.meta.user_overrides = proposed_overrides;
-        record.meta.updated_at = now_ms();
-
-        let write_revision = self
-            .cas_put_record_in_namespace(TOOLS_NAMESPACE, id, &mut record, expected_revision)
-            .await?;
-
-        if let Err(error) = manager
-            .apply_locked()
-            .await
-            .map(|_| ())
-            .map_err(map_runtime_error)
-        {
-            self.emit_audit_apply_failed_in_namespace(
-                TOOLS_NAMESPACE,
-                id,
-                "overrides",
-                Some(before.clone()),
-                None,
-                error.to_string(),
-                headers,
-            )
-            .await;
-            self.rollback_to_raw_after_revision_in_namespace(
-                TOOLS_NAMESPACE,
-                id,
-                raw,
-                write_revision,
-            )
-            .await?;
-            return Err(error);
-        }
-
-        let after_spec = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let after = serde_json::to_value(&after_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        self.emit_audit_with_suffix_in_namespace(
-            AuditAction::Update,
-            TOOLS_NAMESPACE,
-            id,
-            "overrides",
-            Some(before),
-            Some(after.clone()),
-            headers,
-        )
-        .await;
-
-        Ok(after)
-    }
-
-    /// DELETE /v1/config/tools/:id/overrides
-    ///
-    /// Clears all user overrides from a Builtin tool record. Returns the
-    /// effective ToolSpec (which is now the bare base spec, no overrides).
-    pub async fn clear_tool_overrides(
-        &self,
-        id: &str,
-        headers: &HeaderMap,
-    ) -> Result<Value, ConfigServiceError> {
-        let manager = self.runtime_manager()?;
-        let _apply_guard = manager.lock_apply().await;
-
-        let raw = self
-            .store
-            .get(TOOLS_NAMESPACE, id)
-            .await?
-            .ok_or_else(|| ConfigServiceError::NotFound(format!("tools/{id}")))?;
-
-        let mut record = ConfigRecord::<ToolSpec>::from_value(raw.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        if matches!(record.meta.source, RecordSource::User) {
-            return Err(overrides_not_supported_for_user_record());
-        }
-
-        // Short-circuit: if overrides are already None, this is a no-op — skip
-        // the store write, apply_locked, and audit emit.
-        if record.meta.user_overrides.is_none() {
-            return serde_json::to_value(&record.spec)
-                .map_err(|e| ConfigServiceError::Serialization(e.to_string()));
-        }
-
-        let expected_revision = record.meta.revision;
-
-        let before_spec = apply_overrides(record.spec.clone(), record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let before = serde_json::to_value(&before_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        record.meta.user_overrides = None;
-        record.meta.updated_at = now_ms();
-
-        let write_revision = self
-            .cas_put_record_in_namespace(TOOLS_NAMESPACE, id, &mut record, expected_revision)
-            .await?;
-
-        let apply_result = manager
-            .apply_locked()
-            .await
-            .map(|_| ())
-            .map_err(map_runtime_error);
-        if let Err(error) = apply_result {
-            self.emit_audit_apply_failed_in_namespace(
-                TOOLS_NAMESPACE,
-                id,
-                "overrides",
-                Some(before.clone()),
-                None,
-                error.to_string(),
-                headers,
-            )
-            .await;
-            self.rollback_to_raw_after_revision_in_namespace(
-                TOOLS_NAMESPACE,
-                id,
-                raw,
-                write_revision,
-            )
-            .await?;
-            return Err(error);
-        }
-
-        let after = serde_json::to_value(&record.spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        self.emit_audit_with_suffix_in_namespace(
-            AuditAction::Update,
-            TOOLS_NAMESPACE,
-            id,
-            "overrides",
-            Some(before),
-            Some(after.clone()),
-            headers,
-        )
-        .await;
-
-        Ok(after)
-    }
-
-    /// DELETE /v1/config/tools/:id/overrides/:field
-    ///
-    /// Removes a single field from the user overrides of a Builtin tool record.
-    /// Returns 400 if `field` is not a recognized ToolSpecPatch field.
-    /// Idempotent: if the field is not present in user_overrides, returns the
-    /// current effective spec without writing to the store or emitting an audit event.
-    pub async fn clear_tool_override_field(
-        &self,
-        id: &str,
-        field: &str,
-        headers: &HeaderMap,
-    ) -> Result<Value, ConfigServiceError> {
-        let manager = self.runtime_manager()?;
-        let _apply_guard = manager.lock_apply().await;
-
-        let raw = self
-            .store
-            .get(TOOLS_NAMESPACE, id)
-            .await?
-            .ok_or_else(|| ConfigServiceError::NotFound(format!("tools/{id}")))?;
-
-        let mut record = ConfigRecord::<ToolSpec>::from_value(raw.clone())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-
-        if matches!(record.meta.source, RecordSource::User) {
-            return Err(overrides_not_supported_for_user_record());
-        }
-
-        let expected_revision = record.meta.revision;
-
-        let before_spec = apply_overrides(record.spec.clone(), record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let before = serde_json::to_value(&before_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        // Validate that field is recognized by ToolSpecPatch before mutating.
-        // Use a null probe: `ToolSpecPatch` accepts null for all Option fields, and
-        // deny_unknown_fields will reject unknown field names.
-        let probe = Value::Object({
-            let mut m = Map::new();
-            m.insert(field.to_string(), Value::Null);
-            m
-        });
-        let _: ToolSpecPatch = serde_json::from_value(probe).map_err(|_| {
-            ConfigServiceError::InvalidPayload(format!("unknown override field: {field}"))
-        })?;
-
-        // Remove the field from existing overrides.
-        let mut existing_map: Map<String, Value> = record
-            .meta
-            .user_overrides
-            .as_ref()
-            .and_then(Value::as_object)
-            .cloned()
-            .unwrap_or_default();
-
-        // Short-circuit: if the field is not present in overrides, this is a no-op —
-        // skip the store write, apply_locked, and audit emit.
-        if !existing_map.contains_key(field) {
-            let effective_spec = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-                .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-            return serde_json::to_value(&effective_spec)
-                .map_err(|e| ConfigServiceError::Serialization(e.to_string()));
-        }
-
-        existing_map.remove(field);
-
-        let merged_value = Value::Object(existing_map.clone());
-        record.meta.user_overrides = if existing_map.is_empty() {
-            None
-        } else {
-            Some(merged_value)
-        };
-        record.meta.updated_at = now_ms();
-
-        let write_revision = self
-            .cas_put_record_in_namespace(TOOLS_NAMESPACE, id, &mut record, expected_revision)
-            .await?;
-
-        let apply_result = manager
-            .apply_locked()
-            .await
-            .map(|_| ())
-            .map_err(map_runtime_error);
-        if let Err(error) = apply_result {
-            self.emit_audit_apply_failed_in_namespace(
-                TOOLS_NAMESPACE,
-                id,
-                &format!("overrides/{field}"),
-                Some(before.clone()),
-                None,
-                error.to_string(),
-                headers,
-            )
-            .await;
-            self.rollback_to_raw_after_revision_in_namespace(
-                TOOLS_NAMESPACE,
-                id,
-                raw,
-                write_revision,
-            )
-            .await?;
-            return Err(error);
-        }
-
-        let after_spec = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-            .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-        let after = serde_json::to_value(&after_spec)
-            .map_err(|e| ConfigServiceError::Serialization(e.to_string()))?;
-
-        self.emit_audit_with_suffix_in_namespace(
-            AuditAction::Update,
-            TOOLS_NAMESPACE,
-            id,
-            &format!("overrides/{field}"),
-            Some(before),
-            Some(after.clone()),
-            headers,
-        )
-        .await;
-
-        Ok(after)
-    }
 }
 
-fn is_nullable_agent_patch_field(field: &str) -> bool {
-    matches!(
-        field,
-        "context_policy" | "allowed_tools" | "excluded_tools" | "reasoning_effort" | "endpoint"
-    )
-}
-
-/// Return the effective spec Value for a stored entry, applying `user_overrides`
-/// when the namespace supports it (currently Agents and tools).
-///
-/// For non-Agent namespaces this is equivalent to `unwrap_spec`.
-fn effective_spec(namespace: ConfigNamespace, value: Value) -> Result<Value, ConfigServiceError> {
-    match namespace {
-        ConfigNamespace::Agents => {
-            let record = ConfigRecord::<AgentSpec>::from_value(value)
-                .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-            let effective = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-                .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-            serde_json::to_value(&effective)
-                .map_err(|e| ConfigServiceError::Serialization(e.to_string()))
-        }
-        _ => Ok(unwrap_spec(value)),
-    }
-}
-
-fn effective_visible_record<T>(value: Value) -> Result<Option<T>, ConfigServiceError>
-where
-    T: serde::de::DeserializeOwned + awaken_contract::ConfigRecordMerge,
-{
-    let record = ConfigRecord::<T>::from_value(value)
-        .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-    if record.meta.hidden {
-        return Ok(None);
-    }
-    apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-        .map(Some)
-        .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))
-}
-
-fn effective_tool_spec(value: Value) -> Result<Value, ConfigServiceError> {
-    let record = ConfigRecord::<ToolSpec>::from_value(value)
-        .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-    let effective = apply_overrides(record.spec, record.meta.user_overrides.as_ref())
-        .map_err(|e| ConfigServiceError::InvalidPayload(e.to_string()))?;
-    serde_json::to_value(&effective).map_err(|e| ConfigServiceError::Serialization(e.to_string()))
-}
-
-fn redact_provider_error(message: &str, spec: &ProviderSpec) -> String {
-    let mut needles = Vec::new();
-    if let Some(api_key) = spec.api_key.as_ref() {
-        let secret = api_key.expose_secret();
-        push_redaction_needle(&mut needles, secret);
-        if let Ok(value) = serde_json::from_str::<Value>(secret) {
-            collect_string_value_needles(&value, &mut needles);
-        }
-    }
-    collect_adapter_option_error_needles(&spec.adapter_options, &mut needles);
-
-    needles.sort_by_key(|needle| std::cmp::Reverse(needle.len()));
-    needles.dedup();
-
-    let mut redacted = message.to_owned();
-    for needle in needles {
-        redacted = redacted.replace(&needle, "***");
-    }
-    redacted
-}
-
-fn collect_adapter_option_error_needles(
-    options: &std::collections::BTreeMap<String, Value>,
-    needles: &mut Vec<String>,
-) {
-    for (key, value) in options {
-        let lower = key.to_lowercase();
-        if lower == "headers" {
-            collect_string_key_and_value_needles(value, needles);
-        } else if lower.contains("api_key")
-            || lower.contains("bearer")
-            || lower.contains("private_key")
-            || lower.contains("token")
-            || lower.contains("password")
-            || lower.contains("secret")
-        {
-            collect_string_value_needles(value, needles);
-        }
-    }
-}
-
-fn collect_string_value_needles(value: &Value, needles: &mut Vec<String>) {
-    match value {
-        Value::String(value) => push_redaction_needle(needles, value),
-        Value::Array(values) => {
-            for value in values {
-                collect_string_value_needles(value, needles);
-            }
-        }
-        Value::Object(values) => {
-            for value in values.values() {
-                collect_string_value_needles(value, needles);
-            }
-        }
-        Value::Null | Value::Bool(_) | Value::Number(_) => {}
-    }
-}
-
-fn collect_string_key_and_value_needles(value: &Value, needles: &mut Vec<String>) {
-    match value {
-        Value::String(value) => push_redaction_needle(needles, value),
-        Value::Array(values) => {
-            for value in values {
-                collect_string_key_and_value_needles(value, needles);
-            }
-        }
-        Value::Object(values) => {
-            for (key, value) in values {
-                push_redaction_needle(needles, key);
-                collect_string_key_and_value_needles(value, needles);
-            }
-        }
-        Value::Null | Value::Bool(_) | Value::Number(_) => {}
-    }
-}
-
-fn push_redaction_needle(needles: &mut Vec<String>, value: &str) {
-    let value = value.trim();
-    if value.len() >= 4 {
-        needles.push(value.to_owned());
-    }
-}
-
-/// Classify a tool's source from its id.
-///
-/// MCP tools registered by `awaken-ext-mcp` follow `mcp__{server}__{tool}`.
-/// The global registry currently holds only built-in tools, but the classifier
-/// is written defensively so it still works if MCP tools are ever surfaced here.
-fn classify_tool_source(id: &str) -> Value {
-    if let Some(rest) = id.strip_prefix("mcp__") {
-        // Extract server id: the segment between the two `__` delimiters.
-        let server = rest.split("__").next().unwrap_or(rest);
-        return json!({ "kind": "mcp", "id": server });
-    }
-    json!({ "kind": "builtin" })
-}
-
-fn map_runtime_error(error: ConfigRuntimeError) -> ConfigServiceError {
+pub(super) fn map_runtime_error(error: ConfigRuntimeError) -> ConfigServiceError {
     match error {
         ConfigRuntimeError::UnsupportedProviderAdapter(_)
         | ConfigRuntimeError::InvalidConfig(_) => {
@@ -2409,23 +855,6 @@ fn map_runtime_error(error: ConfigRuntimeError) -> ConfigServiceError {
         | ConfigRuntimeError::ChangeListener(_) => ConfigServiceError::Apply(error.to_string()),
         ConfigRuntimeError::Storage(error) => ConfigServiceError::Storage(error),
     }
-}
-
-fn into_object(value: Value) -> Result<Map<String, Value>, ConfigServiceError> {
-    match value {
-        Value::Object(object) => Ok(object),
-        _ => Err(ConfigServiceError::InvalidPayload(
-            "expected JSON object body".into(),
-        )),
-    }
-}
-
-fn from_value<T>(value: &Value) -> Result<T, ConfigServiceError>
-where
-    T: serde::de::DeserializeOwned,
-{
-    serde_json::from_value(value.clone())
-        .map_err(|error| ConfigServiceError::InvalidPayload(error.to_string()))
 }
 
 #[cfg(test)]
